fix: :bug: reverting to previous style. of deployment for db backup file...

+++ app/data/scripts/update_all_data.py
@@ -3,7 +3,7 @@ import os
 from pathlib import Path
 from zipfile import ZIP_DEFLATED, ZipFile
 
-from app.core.config import BLOCKS_JSON, CHARACTERS_JSON, DB_FILE, JSON_FOLDER, PLANES_JSON, S3_BUCKET_URL
+from app.core.config import BLOCKS_JSON, CHARACTERS_JSON, DB_FILE, DB_FOLDER, JSON_FOLDER, PLANES_JSON, S3_BUCKET_URL
 from app.core.result import Result
 from app.core.util import run_command
 from app.data.scripts.get_xml_unicode_db import get_xml_unicode_database
@@ -14,10 +14,12 @@ from app.data.scripts.util import finish_task, start_task
 
 def update_all_data(version: str):
     spinner = start_task(f"Downloading Unicode XML Database v{version} from unicode.org...")
+    spinner.stop_and_persist()
     get_xml_result = get_xml_unicode_database(version)
     if get_xml_result.failure:
         finish_task(spinner, False, "Download failed! Please check the internet connection.")
         return get_xml_result
+    spinner.start()
     finish_task(spinner, True, f"Successfully downloaded Unicode XML Database v{version}!")
     xml_file = get_xml_result.value
 
@@ -33,7 +35,8 @@ def update_all_data(version: str):
             if populate_db_result.failure:
                 return populate_db_result
             if os.environ.get("ENV") == "PROD":
-                xml_file.unlink()
+                if xml_file.exists():
+                    xml_file.unlink()
                 delete_unicode_json_files()
             else:
                 zip_file = backup_sqlite_db()
@@ -60,7 +63,7 @@ def delete_unicode_json_files():
 
 def backup_sqlite_db():
     spinner = start_task("Creating compressed backup file of SQLite database...")
-    zip_file = JSON_FOLDER.joinpath("unicode-api.db.zip")
+    zip_file = DB_FOLDER.joinpath("unicode-api.db.zip")
     with ZipFile(zip_file, "w", ZIP_DEFLATED) as zip:
         zip.write(DB_FILE, f"{DB_FILE. name}")
     finish_task(spinner, True, "Successfully created compressed backup file of SQLite database!")
